qt6-port: fork全量适配Qt 6.8.3编译通过(基于上游qt6分支合并)
- 合并upstream/qt6(PR#767,35提交): QMatrix→QTransform/音频双版本/事件兼容/USE_QT6开关(默认ON)
- 残余QMatrix(18文件)统一转QTransform,相机拖拽补偿逻辑保持SkMatrix不变
- Qt6删除API修复: QDesktopWidget×3/setMargin×14/QTextCodec setCodec×4/QRegExp死include/
  splitRef/qBound歧义/qsizetype收缩/enterEvent签名×5/QWindowsWindowFunctions守卫
- MSVC独有: ewritestream int64_t与long long重载撞车(Linux上游未踩到)
- GDI字体名fromLocal8Bit→fromWCharArray(顺带修非ASCII字体名GBK乱码)
- Windows依赖隔离: ffmpeg头文件复制到sdk/ffmpeg-win(防sdk/include的Qt5头遮蔽Qt6)
- QScintilla源码构建Qt6版(qscintilla2_qt6)装到sdk/qscintilla-qt6,CMake按QT_VERSION_MAJOR分路
- zlib零改动: Qt6.8自带QtZlib头(Z_PREFIX映射z_符号由Qt6Core导出)

+++ src/app/renderhandler.cpp
@@ -632,7 +632,7 @@ int RenderHandler::maxInFlightFrames() const {
     const qint64 w = qint64(mCurrentScene->getCanvasWidth()*resFrac);
     const qint64 h = qint64(mCurrentScene->getCanvasHeight()*resFrac);
     const qint64 frameBytes = qMax<qint64>(1, w*h*4);
-    return int(qBound<qint64>(2, budgetBytes/frameBytes, 4));
+    return int(qBound<qint64>(qint64(2), budgetBytes/frameBytes, qint64(4)));
 }
 
 void RenderHandler::onSceneFrameCached() {
